+++ src/service/dump/abstractDumpService.ts
@@ -3,14 +3,14 @@ import * as vscode from "vscode";
 
 export abstract class AbstractDumpService {
 
-    public dump(node: Node) {
+    public dump(node: Node, withData: boolean) {
         vscode.window.showOpenDialog({ canSelectMany: false, openLabel: "Select export file path", canSelectFiles: false, canSelectFolders: true }).then((folderPath) => {
             if (folderPath) {
-                this.dumpData(node, folderPath[0].fsPath)
+                this.dumpData(node, folderPath[0].fsPath, withData)
             }
         });
     }
 
-    protected abstract dumpData(node: Node, exportPath: string): void;
+    protected abstract dumpData(node: Node, exportPath: string, withData: boolean): void;
 
 }